backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.350
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3

